restructure well-known header defs
