+++ Package.swift
@@ -616,7 +616,6 @@ let package = Package(
         ),
         
         
-        
         // MARK: Test targets
         
         .testTarget(
